requirements.txt
pytorch_lightning==1.4.0
torch==1.7.1
torchvision==0.8.2
transformers==4.6.0
Pillow==8.1.0
tqdm==4.56.0
ipdb==0.13.4
numpy==1.19.5
einops==0.3.0
pyarrow
sacred==0.8.2
pandas==1.1.5
# timm==0.4.12
timm==0.3.2
ftfy
pycocoevalcap
pycocotools
webdataset
nltk
huggingface_hub
